refactor(games): refactoring [SCRUM-201]
